app/requirements.txt
Flask
requests
google-api-core
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
python-dotenv
PyPDF2
python-docx
pymysql
PyJWT
scrypt
langchain
cryptography
langchain-community
langchain-google-genai
faiss-cpu